src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioschedule==0.5.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
click==8.3.0
exceptiongroup==1.3.0
fastapi
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
loguru==0.7.3
multidict==6.6.4
propcache==0.3.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
uvicorn==0.36.0
yarl==1.20.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioschedule==0.5.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.14.2
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
loguru==0.7.3
multidict==6.6.4
propcache==0.3.2
PyJWT==2.10.1
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
uvicorn==0.36.0
yarl==1.20.1
python-json-logger==4.0.0
pydantic[email]
passlib==1.7.4